pyproject.toml
[project]
name = "cv-generator"
version = "1.0.0"
description = "Generate PDF CVs from YAML data"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "weasyprint>=60.0",
]

[project.scripts]
cv-generate = "cv_generator.generator:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.1",
]
